pyproject.toml
[tool.poetry]
name = "llm-resume-builder"
version = "0.1.0"
description = ""
authors = ["yuki_fujii"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.2.2"
numpy = "^1.26.4"
markdown = "^3.6"
black = "^24.4.2"
weasyprint = "^62.0"
pyro-ppl = "^1.9.1"
omegaconf = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
